requirements.txt
Django==1.9.4
django-crispy-forms==1.6.0
docutils==0.12
Pillow==3.1.1
